% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpw}
\alias{bpw}
\title{Bananas per week}
\format{
A data frame with 1000 observations and 2 variables:
\describe{
  \item{id}{Integer. Personal IDs for supermarkt visitors (1--1000).}
  \item{n}{Integer. Average umber of bananas eaten per week (0--10).}
}
}
\usage{
bpw
}
\description{
A fictional dataset containing the average number of bananas eaten per week for 1,000 supermarket visitors.
}
\keyword{datasets}
