% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{create_aliases}
\alias{create_aliases}
\title{Create aliases for food names in the yummm market}
\usage{
create_aliases(food, alias)
}
\arguments{
\item{food}{Character. Food as stored in the yummm market.}

\item{alias}{Character vector. Alias(es) of food.}
}
\description{
Create alias(es) for food names in the yummm market, so that the correct color is printed when a user calls an alias instead of the name under which the food is stored in the yummm market. NB: This function is only used internally.
}
