% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{fill_market}
\alias{fill_market}
\title{Fill the yummm market with food palettes}
\usage{
fill_market(food)
}
\arguments{
\item{food}{Character. Food to be added to the market.}
}
\description{
Fill the yummm market with a 10-color palette extracted from an image of the food. NB: This function is only used internally.
}
