% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yummm.R
\name{in.yummm}
\alias{in.yummm}
\title{Is this food item part of yummm?}
\usage{
in.yummm(food)
}
\arguments{
\item{food}{a character representing a food item of your choice.}
}
\value{
Logical, TRUE or FALSE
}
\description{
Find out whether your favourite food item is part of the `yummm` package.
}
\examples{
in.yummm("banana")

}
