% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yummm.R
\name{in_yummm}
\alias{in_yummm}
\title{Is this food part of yummm?}
\usage{
in_yummm(food)
}
\arguments{
\item{food}{Character vector. Food of your choice.}
}
\description{
Find out whether your favorite food is part of the `yummm` package.
}
\examples{
in_yummm("banana")

}
