% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\docType{data}
\name{yummm.market}
\alias{yummm.market}
\alias{food.items}
\title{The yummm market: a table of food and colours}
\format{Dataframe of character vertors. Column entries: \describe{
  \item{Food}{Food.}
  \item{Colour}{Colours.}
}}
\usage{
food.items
}
\description{
The yummm market: a table of food and colours
}
\keyword{datasets}
