% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{yummm.palette}
\alias{yummm.palette}
\title{Visualize the color palette of the food of your choice}
\usage{
yummm.palette(food)
}
\arguments{
\item{food}{a character vector representing the food of your choice.}
}
\value{
A character string (in hexadecimal format) corresponding to the food of your choice.
}
\description{
Visualization of the 10-color palette of your favorite food.
}
\examples{
yummm.palette("banana")

}
\seealso{
\code{\link{in.yummm}}
}
