% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yummm.R
\name{yummm}
\alias{yummm}
\title{Coloring with all delicious food items}
\usage{
yummm(food)
}
\arguments{
\item{food}{a character representing a food item of your choice.}
}
\value{
The hexadecimal color code corresponding to the food item in `food`.
}
\description{
This is the main function from the `yummm` package.
}
\examples{
yummm("banana")

}
\seealso{
[yummm::in.yummm()]
}
