% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yummm_aliases}
\alias{yummm_aliases}
\title{Aliases for food names}
\format{
A list of food names, each item is a character vector of aliases.
}
\usage{
yummm_aliases
}
\description{
A list with aliases for food names, so that the correct color is printed when a user calls an alias instead of the name under which the food is stored in the yummm market.
}
\keyword{datasets}
