% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{yummm_palette}
\alias{yummm_palette}
\title{Visualize the color palette of the food of your choice}
\usage{
yummm_palette(food, load_fonts = FALSE)
}
\arguments{
\item{food}{Character vector. Food of your choice.}

\item{load_fonts}{TRUE or FALSE. Load fonts to be used in visualisation. Default: FALSE.}
}
\value{
A character string (in hexadecimal format) corresponding to the food of your choice.
}
\description{
Visualization of the 10-color palette of your favorite food.
}
\examples{
yummm_palette("banana")

}
\seealso{
\code{\link{in.yummm}}
}
