% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{fill_market}
\alias{fill_market}
\title{Fill the yummm market with food palettes}
\format{
Dataframe of two columns: \describe{
  \item{Food}{a character. Food.}
  \item{Palette}{a list. Palette of colors in hexadecimal format.}
}
}
\usage{
fill_market(food, n = 10)
}
\description{
Fill the yummm market with a 10-color palette extracted from an image of the food.
}
