% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothie.R
\name{smoothie}
\alias{smoothie}
\title{Delicious smoothies to decorate your graphics}
\usage{
smoothie(
  food,
  shade = rep("01", length(food)),
  alpha = rep(1/length(food), length(food))
)
}
\arguments{
\item{food}{Character vector. Ingredients of your smoothie.}

\item{shade}{Character vector. Shade of your food of choice. Possible id's: "01":"10". Default is set to "01".}

\item{alpha}{Numeric vector. Opacity levels of ingredients of your smoothie. Each value in [0,1]. Default is \code{1/length(food)}.}
}
\value{
A character string (in hexadecimal format) corresponding to the smoothie created with the food of your choice.
}
\description{
Use your favorite smoothie to color your graphics and texts.
}
\examples{
smoothie(c("banana", "kiwi"))

smoothie(c("banana", "kiwi"), alpha=c(0.4, 0.6))

}
\seealso{
\code{\link{yummm}}
}
