% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothie.R
\name{smoothie}
\alias{smoothie}
\title{Delicious smoothies to decorate your graphics}
\usage{
smoothie(food, alpha = rep(1/length(food), length(food)))
}
\arguments{
\item{food}{a character vector representing the ingredients of your smoothie.}

\item{alpha}{a vector opacity levels, with compatible dimensions to \code{foods}. Each level in [0,1]. Default is \code{1/length(food)}.}
}
\value{
A character string (in hexadecimal format) corresponding to the smoothie created with the food of your choice.
}
\description{
Use your favorite smoothie to color your graphics and texts.
}
\examples{
smoothie(c("banana", "kiwi"))

}
\seealso{
\code{\link{yummm}}
}
