% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\docType{data}
\name{yummm.market}
\alias{yummm.market}
\title{The yummm market: a table of food and colors}
\format{Dataframe of character vertors. Column entries: \describe{
  \item{Food}{Food.}
  \item{Color}{Colors.}
}}
\usage{
yummm.market
}
\description{
A table with all food and colors in \code{yummm}. Food and colors are stored in \code{food-colors.R}.
}
\keyword{datasets}
