% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\docType{data}
\name{yummm.market}
\alias{yummm.market}
\title{The yummm market: a table of food and colours}
\format{Dataframe of character vertors. Column entries: \describe{
  \item{Food}{Food.}
  \item{Colour}{Colours.}
}}
\usage{
yummm.market
}
\description{
A table with all food and colours in \code{yummm}. Food and colours are stored in \code{food-colours.R}.
}
\keyword{datasets}
