% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\docType{data}
\name{yummm.market}
\alias{yummm.market}
\alias{food.items}
\title{The Yummm Market: a table of food items and colours}
\format{Dataframe of character vertors. Column entries: \describe{
  \item{Food}{Food items.}
  \item{Colour}{Colours.}
}}
\usage{
food.items
}
\description{
The Yummm Market: a table of food items and colours
}
\keyword{datasets}
