% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yummm.R
\name{yummm}
\alias{yummm}
\title{Delicious colors to decorate your graphics}
\usage{
yummm(food)
}
\arguments{
\item{food}{a character representing a food item of your choice.}
}
\value{
A character string (in hexadecimal format) corresponding to the food item of your choice.
}
\description{
Use your favourite food items to colour your graphics and texts.
}
\examples{
yummm("banana")

}
\seealso{
\code{\link{in.yummm}}
}
