% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yummm_market}
\alias{yummm_market}
\title{The yummm market: a table of food and palettes}
\format{
Dataframe of two columns: \describe{
  \item{Food}{a character. Food.}
  \item{Palette}{a list. Palette of colors in hexadecimal format.}
}
}
\usage{
yummm_market
}
\description{
A table with all food and palettes in \code{yummm}. The market is filled using \code{fill_market}.
}
\keyword{datasets}
